requirements.txt
cffi==1.2.1
cryptography==1.0.2
enum34==1.0.4
gspread==0.2.5
hjson==1.5.2
httplib2==0.9.2
idna==2.0
ipaddress==1.0.14
oauth2client==1.5.1
pyasn1==0.1.9
pyasn1-modules==0.0.8
pycparser==2.14
pyes==0.99.5
pyOpenSSL==0.15.1
python-dateutil==2.4.2
pytz==2015.6
requests==2.8.0
rsa==3.2
six==1.10.0
urllib3==1.12
wheel==0.24.0
bugzilla=1.0.0
